third_party/blink/web_tests/external/wpt/html/browsers/history/the-history-interface/history_reload_referrer-expected.txt

This is a testharness.js-based test.
[FAIL] Step 2: Checking HTTP referrer (is "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub")
  assert_equals: expected "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-1.html" but got "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub"
[FAIL] Step 2: Checking document.referrer (is "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub")
  assert_equals: expected "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-1.html" but got "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub"
[FAIL] Step 3: Checking HTTP referrer (is "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub")
  assert_equals: expected "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-1.html" but got "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub"
[FAIL] Step 3: Checking document.referrer (is "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub")
  assert_equals: expected "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-1.html" but got "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub"
[FAIL] Step 4: Checking HTTP referrer (is "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub")
  assert_equals: expected "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-1.html" but got "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub"
[FAIL] Step 4: Checking document.referrer (is "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub")
  assert_equals: expected "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-1.html" but got "http://web-platform.test:8001/html/browsers/history/the-history-interface/history_reload_referrer-2.html?pipe=sub"
Harness: the test ran to completion.
